chart.js 데이터 연동 완료 main.jsp boxoffice 추가 테스트 및 detail페이지와 연동 완료

+++ src/com/batcha/starsData/model/starsDataDAO.java
@@ -86,11 +86,11 @@ public class starsDataDAO {
 	
 	//셀렉트인데.. 평점 한개만 가져오는 것
 	//이 메소드 결과값이 0이면 결국 평점을 입력하지 않았다는 뜻인데
-	public float getStarsByMemNo(int memNo, int mvNo) throws SQLException {//평점을 부여한 경우 몇점인지 찾기
+	public int getStarsByMemNo(int memNo, int mvNo) throws SQLException {//평점을 부여한 경우 몇점인지 찾기
 		Connection con=null;
 		PreparedStatement ps=null;
 		ResultSet rs=null;
-		float memStars=0;
+		int memStars=0;
 		
 		try {
 			con=pool.getConnection();
